+++ packages/blogpress/CHANGELOG.md
@@ -1,5 +1,12 @@
 # blogpress
 
+## 2.0.1-beta.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @sugarat/theme@0.1.36-beta.2
+
 ## 2.0.1-beta.0
 
 ### Patch Changes